Uncommented the two lines againa and refactored location of GlobalStateProvider.js
