+++ panda_kiosk/src/components/GlobalStateProvider.js
